+++ R/prune.R
@@ -22,22 +22,22 @@
 #' @template example-prune
 rm_data <- function(
   root = ".", path = NULL, recursive = TRUE, ...
-){
+) {
   UseMethod("rm_data", root)
 }
 
 #' @export
 rm_data.default <- function(
   root, path = NULL, recursive = TRUE, ...
-){
-  stop("a 'root' of class ", class(root), " is not supported")
+) {
+  stop("a 'root' of class ", class(root), " is not supported", call. = FALSE)
 }
 
 #' @export
 #' @importFrom assertthat assert_that is.flag
 rm_data.character <- function(
   root = ".", path = NULL, recursive = TRUE, ...
-){
+) {
   to_do <- list_data(root = root, path = path, recursive = recursive)
   if (length(to_do) == 0) {
     return(to_do)
@@ -63,7 +63,7 @@ rm_data.character <- function(
 rm_data.git_repository <- function(
   root, path = NULL, recursive = TRUE, ..., stage = FALSE,
   type = c("unmodified", "modified", "ignored", "all")
-){
+) {
   type <- match.arg(type)
   to_do <- list_data(root = root, path = path, recursive = recursive)
   if (length(to_do) == 0) {
@@ -116,22 +116,22 @@ rm_data.git_repository <- function(
 #' @template example-prune
 prune_meta <- function(
   root = ".", path = NULL, recursive = TRUE, ...
-){
+) {
   UseMethod("prune_meta", root)
 }
 
 #' @export
 prune_meta.default <- function(
   root, path = NULL, recursive = TRUE, ...
-){
-  stop("a 'root' of class ", class(root), " is not supported")
+) {
+  stop("a 'root' of class ", class(root), " is not supported", call. = FALSE)
 }
 
 #' @export
 #' @importFrom assertthat assert_that is.flag noNA
 prune_meta.character <- function(
   root = ".", path = NULL, recursive = TRUE, ...
-){
+) {
   assert_that(is.string(root))
   root <- normalizePath(root, winslash = "/", mustWork = TRUE)
   assert_that(is.string(path))
@@ -153,7 +153,7 @@ prune_meta.character <- function(
                   FUN.VALUE = NA, root = root, message = "none")
   if (any(!check)) {
     warning("Invalid metadata files found. See ?is_git2rmeta():\n",
-            paste(to_do_base[!check], collapse = "\n"))
+            paste(to_do_base[!check], collapse = "\n"), call. = FALSE)
   }
   to_do <- to_do[check]
 
@@ -171,7 +171,7 @@ prune_meta.character <- function(
 #' @rdname prune_meta
 prune_meta.git_repository <- function(
   root, path = NULL, recursive = TRUE, ..., stage = FALSE
-){
+) {
   root_wd <- normalizePath(workdir(root), winslash = "/")
   assert_that(is.string(path))
   path <- file.path(root_wd, path)
@@ -207,6 +207,7 @@ prune_meta.git_repository <- function(
     changed <- gsub("\\.tsv$", ".yml", file.path(root_wd, changed, fsep = "/"))
     if (any(to_do %in% changed)) {
       stop(
+        call. = FALSE,
 "cannot remove and stage metadata in combination with removed but unstaged data"
       )
     }
@@ -216,7 +217,8 @@ prune_meta.git_repository <- function(
     ))
     changed <- gsub("\\.tsv$", ".yml", file.path(root_wd, changed, fsep = "/"))
     if (any(to_do %in% changed)) {
-      warning("data removed and staged, metadata removed but unstaged")
+      warning("data removed and staged, metadata removed but unstaged",
+              call. = FALSE)
     }
   }
   file.remove(to_do)