+++ R/relabel.R
@@ -69,7 +69,8 @@ relabel <- function(file, root = ".", change) {
 
 #' @export
 relabel.default <- function(file, root, change) {
-  stop("a 'change' of class ", class(change), " is not supported")
+  stop("a 'change' of class ", class(change), " is not supported",
+       call. = FALSE)
 }
 
 #' @export
@@ -94,7 +95,7 @@ relabel.list <- function(file, root = ".", change) {
   optimize <- meta_data[["..generic"]][["optimize"]]
   if (!optimize) {
     stop("relabelling factors on verbose data leads to large diffs.
-Use write_vc() instead.")
+Use write_vc() instead.", call. = FALSE)
   }
   assert_that(
     all(names(change) %in% names(meta_data)),
@@ -130,7 +131,7 @@ Use write_vc() instead.")
 #' @export
 #' @importFrom assertthat assert_that has_name
 #' @importFrom stats setNames
-relabel.data.frame <- function(file, root, change) {
+relabel.data.frame <- function(file, root, change) { #nolint
   assert_that(
     has_name(change, "factor"),
     has_name(change, "old"),