+++ NEWS.md
@@ -1,3 +1,26 @@
+git2rdata 0.2.0 (2019-11-08)
+=================================
+
+### BREAKING FEATURES
+
+  * Calculation of data hash has changed (#53). 
+    You must use `upgrade_data()` to read data stored by an older version.
+  * `is_git2rdata()` and `upgrade_data()` do not test equality in data hashes any more (but `read_vc()` still does).
+  * `write_vc()` and `read_vc()` fail when `file` is a location outside of `root` (#50).
+  * Reordering factor levels requires `strict = TRUE`.
+
+### Bugfixes
+
+  * Linux and Windows machines now generated the same data hash (#49).
+
+### NEW FEATURES
+
+  * Internal sorting uses the "C" locale, regardless of the current locale.
+  * `read_vc()` reads older stored in an older version (#44). 
+    When the version is too old, it prompts to `upgrade_data()`.
+  * Improve `warnings()` and `error()` messages.
+  * Use vector version of logo.
+
 git2rdata 0.1 (2019-06-04)
 ============================
 
@@ -23,7 +46,8 @@ git2rdata 0.0.4 (2019-05-16)
   * The meta data gains a data hash. A mismatch throws a warning when reading the object. This tolerates updating the data by other software, while informing the user that such change occurred.
   * `is_git2rmeta()` validates metadata.
   * `list_data()` lists files with valid metadata. 
-  * `rm_data()` and `prune_meta()` remove files with valid metadata. Other files are untouched.
+  * `rm_data()` and `prune_meta()` remove files with valid metadata. 
+    They don't touch `tsv` file without metadata or `yml` files not associated with `git2rdata`.
   *  Files with invalid metadata yield a warning with `list_data()`, `rm_data()` and `prune_meta()`.
   
 ### Bugfixes